feat: use a fixed word to check if remap happened
we simply use the first word after the IVT, because its location is
fixed. We can read and write to it even if the remap has not happened
yet. Problem was, that before a remap, one cant check the memory at 0x0
as that seems to lead to some error. So we need this flag to save if WE
have performed a remap yet and to avoid remapping again (resetting the
remap)